dpo-data-generator/pyproject.toml
[project]
name = "dpo-data-generator"
version = "0.1.0"
description = "DPO 训练数据生成工具 - 用于日常对话场景收集偏好数据"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
]

[project.scripts]
dpo-gen = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "I"]
